optimized city screen for smart phone - moved "buy" button to city info table - removed "add to queue" button - expand icon changed to android defaults - CityStatsTable: big scrollable area, expandable
